parallel_agents/pyproject.toml
[project]
name = "parallel-agents"
version = "0.1.0"
description = "Parallel research agents orchestrator"
authors = [{ name = "Terence Liu", email = "llebox@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai-agents>=0.2.0",
  "openai>=1.30.0",
  "tenacity>=8.2.3",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "numpy>=1.26.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
parallel-agents = "src.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
